backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-multipart
python-dotenv
Pillow
google-genai
